Supplement priority with our own counter
The problem with sorting by priority is that the order of two items with
the same priority is undefined, leading to unexpected behavior. This
gets around that by adding a fractional counter to the priority set; so
that if two speed bumps are registered at priority 10, the one first
registered will be processed first.

Yes, this will have some unexpected behavior if a user registers more
than 100000 speed bumps.  But at that point I think they'd have more
than one problem.