Cache the result of inserting speed bumps
Adds the result of speed bump insertion over a content string to the
object cache. The cache key is generated from the content string and the
speed bumps registered to insert.

Allows filtering of the cache key content, in the case that the
result is dependant on other factors than those two variables (meta
values is one example, or template-dependant factors).